Lift dds image conversion into a separate 'gpu' window/process
This creates a new global/persistent BrowserWindow (GPUWindow) that is always in the background.
The app window can make an ipc call to the main process requesting a dds conversion.
The main process then makes an ipc call to the GPUWindow which does the same dds conversion code that existed prior to this commit.
It then makes a response ipc call (`gpu-dataURL`) that passes back the contents and the converted dataURL.
The AppWindow listens for that ipc and then forwards it to the ui.
The UI then checks to see if the contents match the contents passed in the request, and if so, resolves the dataURL (this is necessary to support multiple conversions happening at the same time).